refactor(common/rest): moves staticmethod utils to base RESTClient
