+++ alpaca/broker/constants.py
@@ -1,4 +1,1 @@
-from typing import TypeVar
-
 BROKER_DOCUMENT_UPLOAD_LIMIT = 10
-PageItem = TypeVar("PageItem")
