+++ alpaca/common/constants.py
@@ -1,3 +1,7 @@
+from typing import TypeVar
+
 DATA_V2_MAX_LIMIT = 10000  # max items per api call
 
 ACCOUNT_ACTIVITIES_DEFAULT_PAGE_SIZE = 100
+
+PageItem = TypeVar("PageItem")  # Generic type for an item from a paginated request.
